--- packages/dashboard/src/constants.ts
@@ -1,1 +1,0 @@
-export const PERSONA_NAME_MAX_LENGTH = 24
