--- packages/dashboard/src/utils.ts
@@ -1,3 +1,0 @@
-export function checkInputLengthExceed(value: string, length: number) {
-    return Array.from(value).length >= length
-}
